// src/apis/models/Volo_Abp_Identity_IdentityRoleCreateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_Abp_Identity_IdentityRoleCreateDto = {
    readonly extraProperties?: Record<string, any> | null;
    name: string;
    isDefault?: boolean;
    isPublic?: boolean;
};